Attempt to factor out concatenate_cubes and associated private functions from the ensemble_calibration_utilities.
