+++ lib/improver/tests/ensemble_calibration/ensemble_calibration/test_ApplyCoefficientsForEnsembleCalibration.py
@@ -47,8 +47,7 @@ import warnings
 
 from improver.ensemble_calibration.ensemble_calibration import (
     ApplyCoefficientsFromEnsembleCalibration as Plugin)
-from improver.ensemble_calibration.ensemble_calibration_utilities import (
-    concatenate_cubes)
+from improver.utilities.cube import concatenate_cubes
 from improver.tests.ensemble_calibration.ensemble_calibration.\
     helper_functions import (set_up_temperature_cube,
                              add_forecast_reference_time_and_forecast_period)
